refactored so all our network activity is in the network_service file and fetchStock is now in the stock_service file. I also moved my Stock and StockList classes in to the appropriate stock.dart file within the models folder
