+++ flutter_module/.android/build.gradle.orig
@@ -1,0 +1,29 @@
+// Generated file. Do not edit.
+<<<<<<< HEAD
+
+apply plugin: 'com.google.gms.google-services'
+=======
+apply plugin: 'com.google.gms.google-services' 
+>>>>>>> 1291f29d61e814922067650ce52d9a88e8557796
+buildscript {
+    repositories {
+        google()
+        jcenter()
+    }
+
+    dependencies {
+        classpath 'com.android.tools.build:gradle:3.5.0'
+        classpath 'com.google.gms:google-services:4.3.3'
+    }
+}
+
+allprojects {
+    repositories {
+        google()
+        jcenter()
+    }
+}
+
+task clean(type: Delete) {
+    delete rootProject.buildDir
+}
