+++ flutter_module/lib/screens/stock_info_screen.dart
@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:fluttermodule/constants.dart';
 import 'package:fluttermodule/services/stock_service.dart';
 import 'package:fluttermodule/components/price_panel.dart';
-import 'package:fluttermodule/api_keys.dart';
+import 'package:fluttermodule/models/stock.dart';
 import 'package:syncfusion_flutter_charts/charts.dart'; // charts
-import 'package:http/http.dart' as http;
 import 'dart:async';
-import 'dart:convert';
+import 'package:flutter/foundation.dart'; // for debugPrint()
+
 
 List globalStockList;
 
@@ -38,8 +38,10 @@ class _StockInfoScreenState extends State<StockInfoScreen> {
   void initState() {
     updateUI(widget.stockData);
     super.initState();
+    StockService stockService = new StockService();
     futureStock =
-        fetchStock(symbol); // populate the object with data from the API
+        stockService.fetchStock(symbol); // populate the object with data from the API
+    //print("futureStock: " + futureStock.toString());
   }
 
   void updateUI(dynamic stockData) {
@@ -211,82 +213,5 @@ class _StockInfoScreenState extends State<StockInfoScreen> {
 
 //    var open = stockData['Time Series (Daily)']['2020-03-23']['1. open'];
 
-Future<StockList> fetchStock(String symbol) async {
-  // this method fetches our data from the API
-  // a future is a Dart class for async operations, a future represents a potential value OR error that will be available at some future time
-  final response = await http.get(
-      'https://www.alphavantage.co/query?function=TIME_SERIES_DAILY&symbol=' +
-          symbol +
-          '&apikey=' +
-          kAlphaStockAPIKey); // network request
-  // await http.get('https://jsonplaceholder.typicode.com/albums/1');
 
-  if (response.statusCode == 200) {
-    // If the server did return a 200 OK response,
-    // then parse the JSON.
-    return StockList.fromJson(json.decode(response.body)); // get the data
-  } else {
-    // If the server did not return a 200 OK response,
-    // then throw an exception.
-    throw Exception('Failed to load stock');
-  }
-}
 
-class Stock {
-  final String date;
-  final String open;
-  final String high;
-  final String low;
-  final String close;
-  final String volume;
-
-  Stock({this.date, this.open, this.high, this.low, this.close, this.volume});
-
-  @override
-  String toString() {
-    return 'Date: ' +
-        this.date +
-        ' Open: ' +
-        this.open +
-        ' High: ' +
-        this.high +
-        ' Low: ' +
-        this.low +
-        ' Close: ' +
-        this.close +
-        ' Volume: ' +
-        this.volume;
-  }
-}
-
-class StockList {
-  // this object is just a list of stocks
-  final List list;
-
-  StockList({this.list});
-
-  factory StockList.fromJson(Map<String, dynamic> json) {
-    // parse the json into data we can use it
-    if (json['Time Series (Daily)'] == null) {
-      throw Exception(
-          "An error has occured. Are you sure you entered a valid stock ticker?");
-    }
-    int size = json['Time Series (Daily)'].keys.toList().length;
-
-    List<Stock> stockList = [];
-
-    for (int i = 0; i < size; i++) {
-      stockList.add(new Stock(
-        date: json['Time Series (Daily)'].keys.toList()[i],
-        open: json['Time Series (Daily)'].values.toList()[i]['1. open'],
-        high: json['Time Series (Daily)'].values.toList()[i]['2. high'],
-        low: json['Time Series (Daily)'].values.toList()[i]['3. low'],
-        close: json['Time Series (Daily)'].values.toList()[i]['4. close'],
-        volume: json['Time Series (Daily)'].values.toList()[i]['5. volume'],
-      ));
-    }
-    return StockList(
-      list: stockList,
-    );
-  }
-}
