+++ flutter_module/lib/services/network_service.dart
@@ -1,7 +1,6 @@
 import 'dart:convert';
-
 import 'package:http/http.dart' as http;
-
+import 'package:fluttermodule/models/stock.dart';
 class NetworkService {
   final String url;
 
@@ -21,4 +20,17 @@ class NetworkService {
       return null;
     }
   }
+
+  Future<StockList> getDataListFormat() async { // getData method in StockList format,
+    // this is useful for the fetchStock method
+    final response = await http.get(url);
+    if (response.statusCode == 200) {
+      return StockList.fromJson(json.decode(response.body));
+    }
+    else {
+      print(response.statusCode);
+      return null;
+    }
+  }
+
 }
